refactor: rework pull_request.closed event handler
