+++ src/config.ts
@@ -1,6 +1,7 @@
 import { HandlerModule } from "./common";
 import StarCreateHandler from "./events/star.created";
 import PullRequestOpenHandler from "./events/pull_request.opened";
+import PullRequestCloseHandler from "./events/pull_request.closed";
 
 export interface Configuration {
   app_name: string;
@@ -9,6 +10,7 @@ export interface Configuration {
 export const Handlers: HandlerModule[] = [
   StarCreateHandler,
   PullRequestOpenHandler,
+  PullRequestCloseHandler,
 ];
 
 export const AppConfig: Configuration = {
